feat(satellite): rename controllers functions and mod to access_keys
