+++ src/libs/satellite/src/assets/storage/store.rs
@@ -14,7 +14,7 @@ use crate::assets::storage::state::{
 use crate::assets::storage::strategy_impls::{StorageAssertions, StorageState, StorageUpload};
 use crate::auth::store::get_config as get_auth_config;
 use crate::certification::strategy_impls::StorageCertificate;
-use crate::controllers::store::get_controllers;
+use crate::access_keys::store::get_access_keys;
 use crate::memory::state::STATE;
 use crate::types::store::{AssertContext, StoreContext};
 use candid::Principal;
@@ -101,7 +101,7 @@ pub fn delete_asset_store(
     collection: &CollectionKey,
     full_path: FullPath,
 ) -> Result<Option<Asset>, String> {
-    let controllers: Controllers = get_controllers();
+    let controllers: Controllers = get_access_keys();
     let config = get_config_store();
 
     let context = StoreContext {
@@ -173,7 +173,7 @@ pub fn list_assets_store(
     collection: &CollectionKey,
     filters: &ListParams,
 ) -> Result<ListResults<AssetNoContent>, String> {
-    let controllers: Controllers = get_controllers();
+    let controllers: Controllers = get_access_keys();
 
     let context = StoreContext {
         caller,
@@ -235,7 +235,7 @@ pub fn get_asset_store(
     collection: &CollectionKey,
     full_path: FullPath,
 ) -> Result<Option<Asset>, String> {
-    let controllers: Controllers = get_controllers();
+    let controllers: Controllers = get_access_keys();
 
     let context = StoreContext {
         caller,
@@ -527,7 +527,7 @@ pub fn delete_filtered_assets_store(
     collection: CollectionKey,
     filters: &ListParams,
 ) -> Result<Vec<Option<Asset>>, String> {
-    let controllers: Controllers = get_controllers();
+    let controllers: Controllers = get_access_keys();
 
     let context = StoreContext {
         caller,
@@ -594,7 +594,7 @@ pub fn set_asset_token_store(
     full_path: &FullPath,
     token: &AssetAccessToken,
 ) -> Result<(), String> {
-    let controllers: Controllers = get_controllers();
+    let controllers: Controllers = get_access_keys();
     let config = get_config_store();
 
     let context = StoreContext {
@@ -658,7 +658,7 @@ fn set_asset_token_impl(
 // ---------------------------------------------------------
 
 pub fn create_batch_store(caller: Principal, init: InitAssetKey) -> Result<BatchId, String> {
-    let controllers: Controllers = get_controllers();
+    let controllers: Controllers = get_access_keys();
     let config = get_config();
 
     secure_create_batch_impl(caller, &controllers, &config, init)
@@ -671,7 +671,7 @@ pub fn create_chunk_store(caller: Principal, chunk: UploadChunk) -> Result<Chunk
 }
 
 pub fn commit_batch_store(caller: Principal, commit_batch: CommitBatch) -> Result<Asset, String> {
-    let controllers: Controllers = get_controllers();
+    let controllers: Controllers = get_access_keys();
     let config = get_config();
 
     let asset = commit_batch_storage(
