--- src/libs/satellite/src/sdk/core/controllers.rs
@@ -1,1 +1,0 @@
-pub use crate::controllers::store::{get_admin_controllers, get_controllers};
