+++ src/libs/satellite/src/sdk/core/mod.rs
@@ -1,4 +1,4 @@
-mod controllers;
+mod access_keys;
 mod db;
 mod hooks;
 mod ic;
@@ -6,7 +6,7 @@ mod logs;
 mod random;
 mod storage;
 
-pub use controllers::*;
+pub use access_keys::*;
 pub use db::*;
 pub use hooks::*;
 pub use ic::*;
