+++ src/sputnik/src/sdk/js/controllers.rs
@@ -1,6 +1,6 @@
 use crate::sdk::js::types::shared::{JsControllers, JsUserId};
 use junobuild_satellite::{
-    get_admin_controllers as get_admin_controllers_sdk, get_controllers as get_controllers_sdk,
+    get_access_keys as get_controllers_sdk, get_admin_access_keys as get_admin_controllers_sdk,
 };
 use junobuild_shared::segments::controllers::{
     controller_can_write as is_controller_sdk, is_admin_controller as is_admin_controller_sdk,
